# backend/app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ChronoSift API",
    "description": "Forensic Timeline Analysis API - Comprehensive documentation for all endpoints",
    "version": "1.0.0",
    "contact": {
      "name": "ChronoSift",
      "url": "https://github.com/yourusername/chronosift"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login"
      }
    },
    "schemas": {
...
      "Comment": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "content": {"type": "string"},
          "user_id": {"type": "integer"},
          "entry_id": {"type": "integer"},
          "parent_id": {"type": "integer", "nullable": true},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
  "paths": {
...
    "/comments/entry/{entry_id}": {
      "get": {
        "tags": ["Comments"],
        "summary": "Get comments for a timeline entry",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "List of comments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/Comment"}
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Comments"],
        "summary": "Add a comment to an entry",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": {"type": "string"},
                  "parent_id": {"type": "integer", "description": "For threaded replies"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Comment"}
              }
            }
          }
        }
      }
    },
...
  }
}